apps/Backend/src/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose
pydantic[email]
python-multipart